+++ src/subscription.ts
@@ -1,4 +1,5 @@
 import { exchange as Exchange } from "./exchange";
+import { ProgramAccountType } from "./types";
 import { PublicKey, Context, KeyedAccountInfo } from "@solana/web3.js";
 import * as anchor from "@project-serum/anchor";
 import * as bs58 from "bs58";
@@ -7,11 +8,6 @@ export interface AccountSubscriptionData<T> {
   key: PublicKey;
   account: T;
   context: Context;
-}
-
-export enum ProgramAccountType {
-  // Only support margin accounts for now.
-  MarginAccount = "MarginAccount",
 }
 
 export function subscribeProgramAccounts<T>(